+++ features/FEATURE_LWIP/lwip-interface/lwip/src/api/lwip_err.c
@@ -40,7 +40,7 @@
 #include "lwip/def.h"
 #include "lwip/sys.h"
 
-#include "lwip/errno.h"
+#include "lwip/lwip_errno.h"
 
 #if !NO_SYS
 /** Table to quickly map an lwIP error (err_t) to a socket error
