+++ features/FEATURE_LWIP/lwip-interface/lwip/src/include/lwip/sockets.h
@@ -46,7 +46,7 @@
 #include "lwip/ip_addr.h"
 #include "lwip/err.h"
 #include "lwip/inet.h"
-#include "lwip/errno.h"
+#include "lwip/lwip_errno.h"
 
 #ifdef __cplusplus
 extern "C" {
