+++ features/FEATURE_LWIP/lwip-interface/lwip/src/include/posix/errno.h
@@ -30,4 +30,4 @@
  *
  */
 
-#include "lwip/errno.h"
+#include "lwip/lwip_errno.h"
